+++ src/basic/___features/product/types.ts
@@ -1,0 +1,6 @@
+import { Product } from "../../../types";
+
+export interface ProductWithUI extends Product {
+  description?: string;
+  isRecommended?: boolean;
+}
